Add CMakeLists to the Nuklear project
Nuklear in itself is not written with CMake or other build systems in
mind. It is not even written for a standard handmade C compiler
compilation: source *.c files are written with the sole purpose of
being mashed together in the final "nuklear.h" header file.

This introduces a bit of friction with a classic C compilation:
the most important one being definitions which should be in a header file
are present in source files, making them not visible to other source/header
files.

With this in mind, I tried without changing the code too much to make
them all behave like good old source/header files. I did not touch
everything the should be touched, I limited myself to the definitions
which were necessary for the current use case of the Sabbie project.